Define Rails metric suites
