+++ lib/librato/rails/railtie.rb
@@ -39,6 +39,19 @@ module Librato
           end
         end
 
+        tracker = config.librato_rails.tracker
+        require_relative 'subscribers/action' if tracker.suite_enabled?(:rails_action)
+        require_relative 'subscribers/cache' if tracker.suite_enabled?(:rails_cache)
+        require_relative 'subscribers/controller' if tracker.suite_enabled?(:rails_controller)
+        require_relative 'subscribers/mail' if tracker.suite_enabled?(:rails_mail)
+        require_relative 'subscribers/method' if tracker.suite_enabled?(:rails_method)
+        require_relative 'subscribers/render' if tracker.suite_enabled?(:rails_render)
+        require_relative 'subscribers/sql' if tracker.suite_enabled?(:rails_sql)
+        require_relative 'subscribers/status' if tracker.suite_enabled?(:rails_status)
+
+        Librato::Rails::VersionSpecifier.supported(min: '4.2') do
+          require_relative 'subscribers/job'if tracker.suite_enabled?(:rails_job)
+        end
       end
 
     end
