+++ lib/librato/rails/subscribers/controller.rb
@@ -4,24 +4,10 @@ module Librato
 
       # Controllers
 
-      def self.watch_controller_action(controller, action)
-        @watches ||= []
-        @watches << "#{controller}##{action}".freeze
-      end
+      ActiveSupport::Notifications.subscribe 'process_action.action_controller' do |*args|
 
-      AS = ActiveSupport
-      AS::Notifications.subscribe 'process_action.action_controller' do |*args|
-
-        event = AS::Notifications::Event.new(*args)
-        controller = event.payload[:controller]
-        action = event.payload[:action]
-
-        format = event.payload[:format] || "all"
-        format = "all" if format == "*/*"
-        status = event.payload[:status]
-        http_method = event.payload[:method]
+        event = ActiveSupport::Notifications::Event.new(*args)
         exception = event.payload[:exception]
-        # page_key = "request.#{controller}.#{action}_#{format}."
 
         collector.group "rails.request" do |r|
 
@@ -35,43 +21,8 @@ module Librato
             r.timing 'time.view', event.payload[:view_runtime] || 0, percentile: 95
           end
 
-          if http_method
-            verb = http_method.to_s.downcase
-            r.group 'method' do |m|
-              m.increment verb
-              m.timing "#{verb}.time", event.duration
-            end
-          end
-
-          unless status.blank?
-            r.group 'status' do |s|
-              s.increment status
-              s.increment "#{status.to_s[0]}xx"
-              s.timing "#{status}.time", event.duration
-              s.timing "#{status.to_s[0]}xx.time", event.duration
-            end
-          end
-
           r.increment 'slow' if event.duration > 200.0
         end # end group
-
-        if @watches && @watches.index("#{controller}##{action}")
-          source = "#{controller}.#{action}.#{format}"
-          collector.group 'rails.action.request' do |r|
-
-            r.increment 'total', source: source
-            r.increment 'slow', source: source if event.duration > 200.0
-            r.timing    'time', event.duration, source: source, percentile: 95
-
-            if exception
-              r.increment 'exceptions', source: source
-            else
-              r.timing 'time.db', event.payload[:db_runtime] || 0, source: source, percentile: 95
-              r.timing 'time.view', event.payload[:view_runtime] || 0, source: source, percentile: 95
-            end
-
-          end
-        end
 
       end # end subscribe
